chromium/third_party/blink/web_tests/wpt_internal/display-lock/css-subtree-visibility/subtree-visibility-075.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Subtree Visibility: accessibility focus</title>
<link rel="author" title="Aaron Leventhal" href="mailto:aleventhal@chromium.org">
<link rel="help" href="https://github.com/WICG/display-locking">
<meta name="assert" content="subtree-visibility auto accessible object subtrees are not destroyed after accessing name">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
.spacer {
  height: 3000px;
}
.auto {
  subtree-visibility: auto;
}
</style>

<div class=spacer></div>
<div id="hidden" class=auto>
  foo
  <div id="child" tabindex="0">
    bar
  </div>
</div>

<script>
function axElementById(id) {
  return accessibilityController.accessibleElementById(id);
}

async_test(async(t) => {
  const hiddenEl = document.getElementById("hidden");
  let axHidden = axElementById("hidden");
  t.step(() => {
    assert_equals(axHidden.childrenCount, 3, "Child count when hidden" );
    axHidden.childAtIndex(1).name;   // Referring to the name should have no effect on the number of children.
    assert_equals(axHidden.childrenCount, 3, "Child count when hidden, and accessing name of first child" );
    t.done();
  });
}, "Accessiblility name calculation on child of subtree-visibility: auto does not destroy tree");

</script>
